requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
dataclasses==0.6
Django==3.2
django-filter==22.1
django-graphql-jwt==0.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
PyJWT==1.7.1
pymongo==3.12.1
pytz==2022.2.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.2.4
text-unidecode==1.3
whitenoise==6.2.0
